+++ tests/fixtures/type/input.js
@@ -1,5 +1,5 @@
 const NAMED_TYPE = gql`
-  query testQuery($one: String) {
+  query testQuery($one: String = "apple") {
     getEntity
   }
 `;
